fix(aot): add (int) casts for C2440 compatibility under use native_types
In AOT mode with 'use native_types':
- Array access (['key']), max(), min(), and ?? operator return php::Variant
- Assignment to php::Int triggers C2440: cannot convert php::Var to php::Int

Fixed 4 files:
- ScrollManager.php: (int) on max/min for scroll calculations
- VNodeRenderer.php: (int) on ['borderColor'] ?? default
- ColorScheme.php: (int) on 20 color property assignments from ['key'] ?? default
- TextTheme.php: (int) on 15 font-size property assignments from ['key'] ?? default

+++ framework/Core/ScrollManager.php
@@ -195,7 +195,7 @@ class ScrollManager
             if ($type === 'track') {
                 $clickOffset = $mouseX - $scrollNode->x - (int)($thumbW / 2);
                 $newScrollLeft = (int)($maxScroll * $clickOffset / max($trackW, 1));
-                $newScrollLeft = max(0, min($maxScroll, $newScrollLeft));
+                $newScrollLeft = (int)max(0, min($maxScroll, $newScrollLeft));
                 $this->applyScrollLeft($scrollNode, $newScrollLeft, true);
             } elseif ($type === 'thumb') {
                 $this->scrollDragTarget = $scrollNode;
@@ -217,7 +217,7 @@ class ScrollManager
             if ($type === 'track') {
                 $clickOffset = $mouseY - $scrollNode->y - (int)($thumbH / 2);
                 $newScrollTop = (int)($maxScroll * $clickOffset / max($trackH, 1));
-                $newScrollTop = max(0, min($maxScroll, $newScrollTop));
+                $newScrollTop = (int)max(0, min($maxScroll, $newScrollTop));
                 $this->applyScrollTop($scrollNode, $newScrollTop, true);
             } elseif ($type === 'thumb') {
                 $this->scrollDragTarget = $scrollNode;
